+++ Sources/UnstoppableDomainsResolution/Helpers/AsyncResolver.swift
@@ -15,48 +15,55 @@ internal class AsyncResolver {
     let asyncGroup = DispatchGroup()
 
     func safeResolve<T>(
-        l1func: @escaping @autoclosure GeneralFunction<T>,
-        l2func: @escaping @autoclosure GeneralFunction<T>
+        listOfFunc: Array<GeneralFunction<T>>
     ) throws -> T {
-        let results = try resolve(l1func: l1func(), l2func: l2func())
+        let results = try resolve(listOfFunc: [listOfFunc[0], listOfFunc[1], listOfFunc[2]])
         return try parseResult(results)
     }
 
     func resolve<T>(
-        l1func: @escaping @autoclosure GeneralFunction<T>,
-        l2func: @escaping @autoclosure GeneralFunction<T>
-    ) throws -> [UNSLocation: AsyncConsumer<T>] {
-        var results: [UNSLocation: AsyncConsumer<T>] = [:]
-        let functions: [UNSLocation: GeneralFunction<T>] = [.layer2: l2func, .layer1: l1func]
-        let queue = DispatchQueue(label: "LayerQueque")
-        functions.forEach { function in
-            self.asyncGroup.enter()
-            DispatchQueue.global().async { [weak self] in
-                guard let self = self else { return }
-                do {
-                    let value = try function.value()
-                    queue.sync {
-                        results[function.key] = (value, nil)
-                    }
-                } catch {
-                    queue.sync {
-                        results[function.key] = (nil, error)
-                    }
-                }
-                self.asyncGroup.leave()
+            listOfFunc: Array<GeneralFunction<T>>
+        ) throws -> [UNSLocation: AsyncConsumer<T>] {
+            var results: [UNSLocation: AsyncConsumer<T>] = [:]
+            var functions: [UNSLocation: GeneralFunction<T>] = [
+                .layer2: listOfFunc[1], .layer1: listOfFunc[0]
+            ]
+            
+            if (listOfFunc.count > 2) {
+                functions[.zlayer] = listOfFunc[2]
             }
+            
+            let queue = DispatchQueue(label: "LayerQueque")
+            functions.forEach { function in
+                self.asyncGroup.enter()
+                DispatchQueue.global().async { [weak self] in
+                    guard let self = self else { return }
+                    do {
+                        let value = try function.value()
+                        queue.sync {
+                            results[function.key] = (value, nil)
+                        }
+                    } catch {
+                        queue.sync {
+                            results[function.key] = (nil, error)
+                        }
+                    }
+                    self.asyncGroup.leave()
+                }
+            }
+            let semaphore = DispatchSemaphore(value: 0)
+            self.asyncGroup.notify(queue: .global()) {
+                semaphore.signal()
+            }
+            semaphore.wait()
+            return results
         }
-        let semaphore = DispatchSemaphore(value: 0)
-        self.asyncGroup.notify(queue: .global()) {
-            semaphore.signal()
-        }
-        semaphore.wait()
-        return results
-    }
+
 
     private func parseResult<T>(_ results: [UNSLocation: AsyncConsumer<T>] ) throws -> T {
-        let l2Result = Utillities.getLayerResultWrapper(from: results, for: .layer2)
-        let l1Result = Utillities.getLayerResultWrapper(from: results, for: .layer1)
+        let l2Result = results[.layer2]!
+        let l1Result = results[.layer1]!
+        let zResult = results[.zlayer]!
 
         if let l2error = l2Result.1 {
             if !isUnregisteredDomain(error: l2error) {
@@ -67,11 +74,22 @@ internal class AsyncResolver {
                 return l2answer
             }
         }
-
+        
         if let l1error = l1Result.1 {
-            throw l1error
+            if !isUnregisteredDomain(error: l1error) {
+                throw l1error
+            }
+        } else {
+            if let l1answer = l1Result.0 {
+                return l1answer
+            }
         }
-        return l1Result.0!
+
+        if let zerror = zResult.1 {
+            throw zerror
+        }
+
+        return zResult.0!
     }
 
     private func isUnregisteredDomain(error: Error?) -> Bool {