+++ CHANGELOG.md
@@ -1,5 +1,13 @@
 # Resolution-swift CHANGELOG
 
+## 5.0.0
+
+- Ability to resolve `.zil` domains from Polygon and Ethereum networks
+- Change default blockchain provider from Infura to Alchemy
+### Breaking changes
+- ZNS support moved under UNS nameserice
+- Remove separate ZNS config
+
 ## 4.0.0
 
 ### Breaking changes